requirements.txt
Flask>=2.2
google-cloud-texttospeech>=2.12.0
python-dotenv>=1.0.0
